pyproject.toml
[project]
name = "goal-dev-spec"
version = "0.1.0"
description = "A goal-driven development specification system using YAML"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.3",
    "platformdirs>=3.5.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "psutil>=5.9.0",
]

[project.scripts]
goal = "goal_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/goal_cli"]